Fixing test for search term in Home tests and moving search term tests to Gallery component so they're not testing Gallery by proxy.
